Fix CodeFactor Issues
